@@ -10,40 +10,41 @@
     "url": "https://github.com/fabnavi/fabnavi5/issues"
   },
   "dependencies": {
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babelify": "^7.3.0",
-    "browserify": "^11.0.1",
     "flux": "^2.1.1",
-    "glob": "^5.0.14",
-    "gulp": "^3.9.0",
-    "gulp-plumber": "^1.0.1",
-    "gulp-sourcemaps": "^1.5.2",
-    "gulp-uglify": "^1.4.0",
-    "gulp-util": "^3.0.6",
-    "jquery": "^2.1.4",
-    "jquery-ujs": "^1.0.4",
+    "jquery": "^3.0.0",
+    "jquery-ujs": "^1.2.1",
     "keymirror": "^0.1.1",
-    "lodash": "^3.10.1",
-    "machina": "^1.1.2",
-    "node-notifier": "^4.2.3",
-    "react": "^0.13.3",
-    "react-jade": "^2.4.0",
-    "react-router": "^0.13.3",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.4.0"
+    "machina": "^2.0.0-1",
+    "react": "^0.14.8",
+    "react-router": "^2.4.0"
+  },
+  "devDependencies": {
+    "react-jade": "git://github.com/edgemethod/react-jade.git#19db22e",
+    "react-jade-loader": "git://github.com/fabnavi/react-jade-loader.git#adb23dc",
+    "webpack": "^1.13.0",
+    "webpack-dev-server": "^1.14.1",
+    "babel-core": "^6.7.7",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-dev-expression": "^0.2.1",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-es2015-loose": "^7.0.0",
+    "babel-preset-es2015-loose-native-modules": "^1.0.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "bower-webpack-plugin": "^0.1.9",
+    "json-loader": "^0.5.4",
+    "node-sass": "^3.6.0",
+    "esdoc": "^0.4.6",
+    "eslint": "^2.5.1"
   },
   "license": "MIT",
   "engines": {
     "node": ">= 0.12"
   },
   "scripts": {
-    "start": "gulp start",
-    "build": "gulp build",
-    "lint": "eslint -c ./.eslintrc app/assets/javascripts/client/*"
-  },
-  "devDependencies": {
-    "eslint": "^2.5.1"
+    "start": "webpack --progress --colors --watch",
+    "build": "webpack",
+    "lint": "eslint -c ./.eslintrc app/assets/javascripts/client/*",
+    "doc": "esdoc -c esdoc.json"
   }
 }
