@@ -1,20 +1,8 @@
-const
-    React = require('react'),
-    EditContent = require('./EditContent.react.js'),
-    EditTitle = require('./EditTitle.react.js'),
-    ProjectListStore = require('../stores/ProjectListStore'),
-    ProjectStore = require('../stores/ProjectStore'),
-    ProjectActionCreator = require('../actions/ProjectActionCreator'),
-    jade = require('react-jade'),
-    Router = require('react-router'),
-    DefaultRoute = Router.DefaultRoute,
-    Link = Router.Link,
-    Route = Router.Route,
-    RouteHandler = Router.RouteHandler,
-
-
-    State = require('../utils/FabnaviStateMachine'),
-    editProject = jade.compileFile(__dirname + '/../templates/EditProject.jade');
+import React from'react';
+import ProjectListStore from'../stores/ProjectListStore';
+import{ Route, RouteHandler, Link, DefaultRoute }from'react-router';
+import State from'../utils/FabnaviStateMachine';
+import editProject from'../templates/EditProject.jade';
 
 const EditProject = React.createClass({
 
@@ -36,10 +24,10 @@ const EditProject = React.createClass({
     this.setState(this.getStateFromStores());
   },
   getInitialState: function(){
-    return  {
-          name : "",
-          description : "",
-        },
+    return {
+      name : "",
+      description : "",
+    },
         this.getStateFromStores();
   },
 
@@ -52,20 +40,20 @@ const EditProject = React.createClass({
 
   getImage: function(){
     console.log("getImage ouou");
-  let project ={};
-  project.content_array=[];
-  project.figure=[];
-  project.figure_id =[];
-  project.project_id = null;
+    let project = {};
+    project.content_array = [];
+    project.figure = [];
+    project.figure_id = [];
+    project.project_id = null;
 
-  for(var i in this.state.projects){
-    if(this.state.projects[i].id == this.context.router.getCurrentParams().projectId){
+    for(var i in this.state.projects){
+    if(this.state.projects[i].id == this.props.params.projectId){
       project.project_id = this.state.projects[i];
       project.content_array = this.state.projects[i].content;
     }
   }
 
-  for(var i in project.content_array){
+    for(var i in project.content_array){
     project.figure.push(project.content_array[i].figure.file.file.thumb.url);
     project.figure_id.push(project.content_array[i].figure.figure_id);
   }
@@ -75,12 +63,12 @@ const EditProject = React.createClass({
   onclick: function(){
     let a = this.getImage();
     console.log("button onclick: " + this.props.id_array);
-    ProjectActionCreator.editContent(a.project_id,this.props.id_array);
+    ProjectActionCreator.editContent(a.project_id, this.props.id_array);
   },
 
 
   delete_num: function(){
-    let num = "DELETE"+String(this.props.id_array.length);
+    let num = "DELETE" + String(this.props.id_array.length);
     return num;
   },
 
@@ -97,11 +85,11 @@ const EditProject = React.createClass({
 
   componentWillMount : function(){
     this.props.id_array = [];
-    ProjectActionCreator.getAllProjects();　
+    ProjectActionCreator.getAllProjects();
   },
 
   componentDidMount : function (){
-    State.transition("pages"); 
+    State.transition("pages");
   },
 
   componentWillUpdate : function(){
