@@ -1,17 +1,17 @@
-const
-    React = require('react'),
-    jade = require('react-jade'),
-    Router = require('react-router'),
-    Route = Router.Route,
-    Link = Router.Link,
-    RouteHandler = Router.RouteHandler,
-    frame = jade.compileFile(__dirname + '/../templates/Frame.jade');
+import React from 'react';
 
-const Frame = React.createClass({
-  contextTypes: {
-    router: React.PropTypes.func
-  },
-  render : frame
-});
+class Frame extends React.Component {
+  constructor(props){
+    super(props);
+  }
 
-module.exports = Frame;
+  render(){
+    return (
+      <div>
+        {this.props.children}
+      </div>
+    );
+  }
+}
+
+export default Frame;
