@@ -167,8 +167,8 @@ const ProjectStore = Object.assign({}, EventEmitter.prototype, {
     console.log("toggletoggle");
     console.log(_project.content.length);
     console.log(_delContent.length);
-    for(let i =0; i<_project.content.length;i++){
-      for(let j =0; j<_delContent.length;j++){
+    for(let i = 0; i < _project.content.length; i++){
+      for(let j = 0; j < _delContent.length; j++){
         if(_project.content[i].figure.figure_id == _delContent[j]){
           console.log(_project.content[i].figure.file.file.thumb.url);
           _project.content[i].figure["_destroy"] = true;
@@ -187,11 +187,11 @@ const ProjectStore = Object.assign({}, EventEmitter.prototype, {
   },
 
   backToHome:function(){
-    if(location.hash.contains("#/manager/detail")){
+    if(location.hash.includes("#/manager/detail")){
       location.hash = "#/manager";
-    }else if(location.hash.contains("#/manager/edit")){
+    } else if(location.hash.includes("#/manager/edit")){
       location.hash = "#/mamager/myprojects";
-    }else{
+    } else {
       location.hash = "#/manager";
     }
   },
@@ -439,7 +439,7 @@ ProjectStore.dispatchToken = AppDispatcher.register(function( action ){
       }, 0);
       setTimeout(function(){
         location.hash = "#/manager/myprojects";
-      },0); 
+      }, 0);
       break;
     case ActionTypes.EDIT_TITLE:
       _project = action.project;
@@ -454,7 +454,7 @@ ProjectStore.dispatchToken = AppDispatcher.register(function( action ){
       }, 0);
       setTimeout(function(){
         location.hash = "#/manager/myprojects";
-      },0);
+      }, 0);
 
       break;
     default :
